extracted loading modal to directory, added to details pages (#221)
